Added support for MLNX OFED 4.9 LTS
Support for all 4.9 LTS releases have been added. This effectively
fixes #2

Note that MLNX_OFED 4.9 LTS seems to be based on `rdma-core` 28.
And the patch is different from newer releases, so it was compared
against the old 28 release. Some change may seem to be missing but
isn't.

It may be relevant to also observe that MLNX_OFED 4.9 is really
dated now and should not be used if not needed. There's a lot of
issues with newer releases of Enterprise Linux 8 and I recoomend to
install MLNX_OFED 4.9 with `--upstream-libs` flag enabled to avoid
yum/dnf breakage.